Extract admin skeletons to _skeleton.tsx files, drop unnecessary async
- Move AdminUsersContentSkeleton and AdminUserDetailSkeleton into
  co-located _skeleton.tsx files (underscore = non-route in App Router)
- loading.tsx and page.tsx now both import from _skeleton.tsx instead
  of loading.tsx importing from page.tsx (two special files coupling)
- AdminUsersPage no longer async — it awaits nothing
- Remove Skeleton import from page files where it is no longer needed
