+++ app/admin/users/[userId]/page.tsx
@@ -20,11 +20,11 @@ import {
 } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Skeleton } from "@/components/ui/skeleton";
 import { adminGetUser, adminGetUserUpgradeRequests } from "@/app/actions/admin";
 import { requireAdmin } from "@/lib/session";
 import { getUserUsage } from "@/lib/usage";
 import { AdminUserActions } from "./AdminUserActions";
+import { AdminUserDetailSkeleton } from "./_skeleton";
 
 async function AdminUserDetailContent({ userId }: { userId: string }) {
   const [session, user, upgradeRequests, usage] = await Promise.all([
@@ -241,76 +241,6 @@ async function AdminUserDetailContent({ userId }: { userId: string }) {
         upgradeRequests={upgradeRequests}
         currentUserId={session.user!.id!}
       />
-    </>
-  );
-}
-
-export function AdminUserDetailSkeleton() {
-  return (
-    <>
-      <div className="flex flex-wrap items-center gap-3">
-        <div className="min-w-0 flex-1 space-y-1.5">
-          <Skeleton className="h-7 w-48" />
-          <Skeleton className="h-4 w-36" />
-        </div>
-        <div className="flex gap-2">
-          <Skeleton className="h-5 w-14 rounded-full" />
-          <Skeleton className="h-5 w-14 rounded-full" />
-        </div>
-      </div>
-
-      <div className="grid gap-6 md:grid-cols-2">
-        <Card>
-          <CardHeader>
-            <Skeleton className="h-5 w-32" />
-          </CardHeader>
-          <CardContent className="space-y-3">
-            <Skeleton className="h-4 w-full" />
-            <Skeleton className="h-4 w-3/4" />
-            <Skeleton className="h-4 w-2/3" />
-            <Skeleton className="h-3 w-40 mt-2" />
-          </CardContent>
-        </Card>
-
-        <Card>
-          <CardHeader>
-            <Skeleton className="h-5 w-24" />
-            <Skeleton className="h-4 w-48 mt-1" />
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div className="grid grid-cols-2 sm:grid-cols-4 gap-4">
-              {[...Array(4)].map((_, i) => (
-                <Skeleton key={i} className="h-20 rounded-lg" />
-              ))}
-            </div>
-            <div className="border-t pt-4 space-y-3">
-              <Skeleton className="h-3 w-36" />
-              {[...Array(2)].map((_, i) => (
-                <div key={i} className="space-y-1">
-                  <div className="flex justify-between">
-                    <Skeleton className="h-4 w-16" />
-                    <Skeleton className="h-4 w-24" />
-                  </div>
-                  <Skeleton className="h-2 w-full rounded-full" />
-                </div>
-              ))}
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-
-      <div className="space-y-4">
-        {[...Array(3)].map((_, i) => (
-          <Card key={i}>
-            <CardHeader>
-              <Skeleton className="h-5 w-40" />
-            </CardHeader>
-            <CardContent>
-              <Skeleton className="h-20 w-full" />
-            </CardContent>
-          </Card>
-        ))}
-      </div>
     </>
   );
 }
