+++ app/admin/users/page.tsx
@@ -10,8 +10,8 @@ import {
 } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Skeleton } from "@/components/ui/skeleton";
 import { adminGetAllUsers, adminGetPendingUpgradeRequests } from "@/app/actions/admin";
+import { AdminUsersContentSkeleton } from "./_skeleton";
 
 const PLAN_COLORS: Record<string, string> = {
   FREE: "secondary",
@@ -203,58 +203,7 @@ async function AdminUsersContent() {
   );
 }
 
-export function AdminUsersContentSkeleton() {
-  return (
-    <>
-      <div className="grid grid-cols-2 gap-4 md:grid-cols-4">
-        {[...Array(4)].map((_, i) => (
-          <Card key={i}>
-            <CardHeader className="flex flex-row items-center justify-between pb-2">
-              <Skeleton className="h-4 w-20" />
-              <Skeleton className="size-4" />
-            </CardHeader>
-            <CardContent>
-              <Skeleton className="h-8 w-12" />
-            </CardContent>
-          </Card>
-        ))}
-      </div>
-      <Card>
-        <CardHeader>
-          <Skeleton className="h-5 w-24" />
-          <Skeleton className="h-4 w-56 mt-1" />
-        </CardHeader>
-        <CardContent>
-          <div className="overflow-x-auto">
-            <div className="min-w-[500px] space-y-0">
-              <div className="flex gap-4 border-b pb-3">
-                <Skeleton className="h-4 w-20" />
-                <Skeleton className="h-4 w-28" />
-                <Skeleton className="h-4 w-14" />
-                <Skeleton className="h-4 w-14" />
-                <Skeleton className="h-4 w-14" />
-                <Skeleton className="h-4 w-10" />
-              </div>
-              {[...Array(6)].map((_, i) => (
-                <div key={i} className="flex items-center gap-4 py-3 border-b last:border-0">
-                  <Skeleton className="h-4 w-28" />
-                  <Skeleton className="h-4 w-36" />
-                  <Skeleton className="h-5 w-16 rounded-full" />
-                  <Skeleton className="h-5 w-14 rounded-full" />
-                  <Skeleton className="h-4 w-8" />
-                  <Skeleton className="h-4 w-8" />
-                  <Skeleton className="h-8 w-16 ml-auto" />
-                </div>
-              ))}
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-    </>
-  );
-}
-
-export default async function AdminUsersPage() {
+export default function AdminUsersPage() {
   return (
     <div className="flex flex-col gap-6">
       <div>
